abortive: gpgme support
There's a lot of language options, with various dependencies associated
with them, such as QT, Node.js, Common Lisp, various Pythons, etc.

I don't want to figure out packaging for all those correctly, for
various use-cases, across OSes and dealing with bugs in integration
across various versions of language runtimes.

I don't use gpgme myself, so the best thing to do is to be honest and
just say "no, not here" instead of try to support something and do a bad
job of it, just creating more noise.

Pushing this commit to show what's needed for the core of integration.