Drop a gpg-wks-client symlink into bin
Rationale is to reflect a planned change and be more convenient.

```
From: Andre Heinecke <aheinecke@gnupg.org>
Subject: Re: GnuPG 2.2 on elder Debian & Ubuntu distros
To: gnupg-devel@gnupg.org
Date: Thu, 21 Mar 2019 14:52:32 +0100

Uhm the libexec thing is on me.
gpg-wks-client and gpg-wks-server were very low level at first and only
intended to be used by higher level applications. So I found libexec to be the
right place.

gpg-wks-client gained more usefulness. So I now think that we should install
it to /usr/bin. I guess we will change that in master for GnuPG 2.3.

On Windows the gpg-wks-client is already installed under bin and in the PATH.
```