Tentative framework for deferred deploys
Need to rework how the os deploy script works, or perhaps be able to
emit directives for later from it, because at present it's too dependent
upon one `$MACHINE` (including guaranteeing its availability to the
`confs/deploy.sh` system) so it's not a simple process.

The goal is to be able to copy all the files across, then invoke aptly
a few times under one gpg-agent, to avoid repeated passphrase prompting,
and then perform AWS CF invalidation just once.