docker: more resilience to env issues
Having now run this on a new machine, little things like "part-way
through long builds, it failed, because there wasn't a docker image
locally for the next machine build" have revealed themselves.

Scan ARGV twice, and on the first pass validate that we have something
we can run for all candidates.  Only if those all succeeded do we
actually do the builds on the second pass; else we give enough details
to make it easy to fix all in one go.

Bad specs will still cause an immediate abort in the first pass, instead
of collecting them all, but that's bad in-repo data not bad env, so I'm
less worried by that.