Bump GnuPG and GnuTLS; handle swdb server down
The versions.gnupg.org server is down, so I can't get an up-to-date
`swdb.lst` file.  I can `touch(1)` it to force it to appear fresh, which I
did, and manually checked all the versions.

Configure using the existing `build_version` overrides support to change
the version of gnupg22 to be 2.2.19 instead of the 2.2.17 in my stale
`swdb.lst` file.

Change the in-Container/VM build flow to handle inability to fetch
current swdb.lst again by copying from the outside env.