keyrings: export again, SHA1 set to weak
I used two exports, one to reflect "the prior world, SHA1 allowed", the
second one with `weak-digest SHA1`.

Whatever the continuing applicability of SHA1 for broader ecosystem
OpenPGP usage might be, in this case I'm working with a small finite set
of keys and I've reached the owners to get them updated as needed, so I
can disable the dependency upon a flawed digest algorithm when verifying
the integrity of security-critical software distributions.