Fix gnutls deps and docker errors
Bail out properly if a build fails, instead of reporting success.

The bump in gmp to 6.2.0 with system headers for 6.1.2 installed exposed
a mismatch in our handling.  The gmp stuff is installing to lib64 not
lib, so handle that install location.

Force remove the GMP dev headers from the system.