For Xenial, install older version of self for Ed25519
The GnuPG 2.2.22 release is signed with an Ed25519 key, which is not supported
by the system GnuPG in Xenial, which made it impossible for us to build.

So now we can bootstrap via having an older version of our own tooling
installed first, to build the newer version.  My assumption here is that at any
given time we won't be adding an OS platform so old that it lacks support for
an algorithm needed for a contemporary GnuPG release, so we'll have a previous
build available.  If that proves false, then effectively we won't have a way to
verify signatures for that OS and we'll need to have a signature bypass,
relying upon having already fetched and verified the signature in a newer OS.