Version patches accepted; Python cleanups
Filter patches to accept _vXYZ_ immediately after the product name and then
filter to only the current version, iff the patch contains that constraint.
Update the README to move a TODO to documented.

I'm not changing the tab indents at this stage, but I've cleaned up enough
other issues for the LSP checker I'm using to drop the noise down to tolerable
levels.

Added :Product typing on params taking those, for better reporting of available
attributes in code editing.